IOS-6324 Moved adapter and protocols into separate folder
